vold: fix 64 bit print error.
Commit "Vold: fix format and scan card related issue" introduced a build
error in a 64 bit environment, where dev_t is no longer of the same size
as an integer.

Fix it by using the bigger printf format, ie. unsigned long, and casting
dev_t. This cast is introduced by the historical discrepancy in dev_t
between 32 bits and 64 bits builds, where dev_t resolves to respectively
uint32_t and uint64_t.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-10321
Category: aosp improvement
Domain: Storage-Removable_Storage
Origin: internal
Upstream-Candidate: no, proprietary
Reviewed-on: https://android.intel.com:443/300548

+++ Volume.cpp
@@ -432,8 +432,8 @@ int Volume::mountVol() {
      * partition, we need to skip this extend partition
      */
     diskdevice = getDiskDevice();
-    sprintf(devicePath, "/dev/block/vold/%d:%d", MAJOR(diskdevice),
-        MINOR(diskdevice));
+    sprintf(devicePath, "/dev/block/vold/%lu:%lu", (unsigned long)MAJOR(diskdevice),
+        (unsigned long)MINOR(diskdevice));
     extendPart = Fat::check_extend(devicePath, n);
 
     for (i = 0; i < n; i++) {
